fix: add Theme
