+++ package.json
@@ -70,12 +70,12 @@
 	},
 	"dependencies": {
 		"@babel/runtime": "^7.17.9",
-		"restify-websocket": "0.0.0-development",
 		"babel-plugin-transform-typescript-metadata": "^0.3.2",
 		"express": "^4.18.1",
 		"js-yaml": "^4.1.0",
 		"lodash": "^4.17.21",
 		"node-pty": "^0.10.1",
+		"restify-websocket": "0.0.0-development",
 		"sqlite3": "^4.0.3",
 		"super-terminal-ui": "0.0.0-development",
 		"typeorm": "0.3.6",
@@ -97,5 +97,4 @@
 			"type": "npm"
 		}
 	}
-
 }
